gharis/gharis-230c91f66cb04f9a94597d4f0c4950a9a1213eb7/backend/requirements.txt
fastapi==0.118.0
uvicorn==0.37.0
pydantic-settings==2.11.0
supabase==2.21.1
python-jose==3.5.0
passlib[bcrypt]==1.7.4
python-multipart
python-dotenv==1.0.1